requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PasteDeploy==1.5.2
psycopg2==2.6.1
py==1.4.29
pyramid==1.5.7
pyramid-jinja2==2.5
pyramid-tm==0.12
pytest==2.7.2
python-dateutil==1.5
repoze.lru==0.6
requests==2.7.0
SQLAlchemy==1.0.5
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.9
WebOb==1.4.1
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
